Moved types into src folder
