+++ frontend/src/components/InGame/Countdown.vue
@@ -12,7 +12,7 @@
 // @ is an alias to /src
 import Vue from 'vue';
 import { mapValues, maxBy } from 'lodash';
-import { Game } from '../../../../types';
+import { Game } from '../../types';
 import store from '../../store';
 
 export default Vue.extend({
