+++ frontend/src/components/InGame/Deck.vue
@@ -27,7 +27,7 @@
 // @ is an alias to /src
 import Vue from 'vue';
 import { includes } from 'lodash';
-import { Player, Round, ResponseCard, Game } from '../../../../types';
+import { Player, Round, ResponseCard, Game } from '../../types';
 import store from '../../store';
 
 export default Vue.extend({
