+++ frontend/src/components/InGame/GameState.vue
@@ -52,7 +52,7 @@
 import Vue from 'vue';
 import Submissions from './Submissions.vue';
 import Countdown from './Countdown.vue';
-import { OtherPlayer, Player, Game, Round } from '../../../../types';
+import { OtherPlayer, Player, Game, Round } from '../../types';
 import store from '../../store';
 
 export default Vue.extend({
