--- frontend/types/index.ts
@@ -1,4 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/triple-slash-reference
-/// <reference path="../../types/index.ts" />
-
-export * from '../../types';
